fixed permissions

+++ commands/moderation/delete.js
@@ -1,10 +1,10 @@
 const { SlashCommandBuilder } = require("@discordjs/builders");
+const { permissions } = require("discord.js");
 
 module.exports = {
     data: new SlashCommandBuilder()
         .setName("delete")
         .setDescription("delete multiple messages")
-        .setDefaultMemberPermissions(4)
         .addIntegerOption((option) =>
             option
                 .setName("number")
@@ -12,17 +12,27 @@ module.exports = {
                 .setRequired(true)
                 .setMaxValue(100)
                 .setMinValue(1)
+        )
+        .addUserOption((option) =>
+            option
+                .setName("target")
+                .setDescription("who's messages do you wanna delete")
+                .setRequired(false)
         ),
     async execute(interaction) {
         let numberOfMessages = interaction.options.getInteger("number");
         let channel = interaction.channel;
         try {
-            await channel.bulkDelete(numberOfMessages).then(() => {
-                interaction.reply(
-                    "`Deleted " + numberOfMessages + " messages`"
-                );
-                setTimeout(() => interaction.deleteReply(), 3000);
-            });
+            if (interaction.member.permissions.has("MANAGE_MESSAGES")) {
+                await channel.bulkDelete(numberOfMessages).then(() => {
+                    interaction.reply(
+                        "`Deleted " + numberOfMessages + " messages`"
+                    );
+                    setTimeout(() => interaction.deleteReply(), 3000);
+                });
+            } else {
+                interaction.reply("you don't have permission to do that");
+            }
         } catch (err) {
             console.log(err);
         }
